requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
fastapi==0.109.0
h11==0.14.0
idna==3.6
mnemonic==0.20
phonenumbers==8.13.29
pycountry==23.12.11
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
requests-sse==0.3.1
sniffio==1.3.0
starlette==0.35.1
stellar-sdk==9.2.0
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
validators==0.22.0
xdrlib3==0.1.1
